Fix initial state and cleanup
