+++ tests/dummy/app/controllers/multiple.js
@@ -22,10 +22,21 @@ export default Ember.Controller.extend({
   dismissOnEsc: true,
   gesturesEnabled: true,
 
-  colors: [
-    'green',
-    'yellow',
-    'blue',
-    'red'
-  ]
+  menus: [{
+    color: 'green',
+    position: 'left',
+    animation: 'slide'
+  }, {
+    color: 'yellow',
+    position: 'right',
+    animation: 'slide'
+  }, {
+    color: 'blue',
+    position: 'left',
+    animation: 'fall-down'
+  }, {
+    color: 'red',
+    position: 'right',
+    animation: 'fall-down'
+  }]
 });
