+++ tests/dummy/app/templates/-header.hbs
@@ -1,0 +1,15 @@
+<div class="header">
+  <img src="images/ember-logo.png" ondragstart="return false;"/>
+  <h1>Burger Menu</h1>
+
+  <div class="description">
+    <p>
+      An off-canvas sidebar component with a collection of animations and styles using CSS transitions.
+    </p>
+  </div>
+
+  <div class="nav">
+    {{#link-to 'index'}}Home{{/link-to}}
+    {{#link-to 'multiple'}}Multiple{{/link-to}}
+  </div>
+</div>
