+++ tests/dummy/app/templates/-inspiration.hbs
@@ -1,0 +1,5 @@
+<div class="inspiration">
+  <p>
+    Inspired by <a href="https://github.com/codrops/SidebarTransitions">Sidebar Transitions</a> by Codrops
+  </p>
+</div>
