+++ tests/dummy/app/templates/-menu.hbs
@@ -1,0 +1,31 @@
+<ul>
+  {{#menu.item}}
+    <a href="https://github.com/offirgolan/ember-burger-menu" target="_blank">
+      <i class="fa fa-github" aria-hidden="true"></i> GitHub
+    </a>
+  {{/menu.item}}
+
+  {{#menu.item}}
+    <a href="https://github.com/offirgolan/ember-burger-menu/issues" target="_blank">
+      <i class="fa fa-bug" aria-hidden="true"></i> Issues
+    </a>
+  {{/menu.item}}
+
+  {{#menu.item}}
+    <a href="https://github.com/offirgolan/ember-burger-menu/pulls" target="_blank">
+      <i class="fa fa-exchange fa-rotate-90" aria-hidden="true"></i> Pull Requests
+    </a>
+  {{/menu.item}}
+
+  {{#menu.item}}
+    <a href="https://github.com/offirgolan/ember-burger-menu/blob/master/README.md" target="_blank">
+      <i class="fa fa-file-text-o" aria-hidden="true"></i> Documentation
+    </a>
+  {{/menu.item}}
+
+  {{#menu.item}}
+    <a href="https://github.com/offirgolan/ember-burger-menu/graphs" target="_blank">
+      <i class="fa fa-bar-chart" aria-hidden="true"></i> Analytics
+    </a>
+  {{/menu.item}}
+</ul>
