+++ tests/dummy/app/templates/index.hbs
@@ -10,37 +10,7 @@
     <a {{action burger.state.actions.close}} class="close {{burger.state.position}} fa fa-times"></a>
     <h2><i class="fa fa-bars" aria-hidden="true"></i> Menu</h2>
 
-    <ul>
-      {{#menu.item}}
-        <a href="https://github.com/offirgolan/ember-burger-menu" target="_blank">
-          <i class="fa fa-github" aria-hidden="true"></i> GitHub
-        </a>
-      {{/menu.item}}
-
-      {{#menu.item}}
-        <a href="https://github.com/offirgolan/ember-burger-menu/issues" target="_blank">
-          <i class="fa fa-bug" aria-hidden="true"></i> Issues
-        </a>
-      {{/menu.item}}
-
-      {{#menu.item}}
-        <a href="https://github.com/offirgolan/ember-burger-menu/pulls" target="_blank">
-          <i class="fa fa-exchange fa-rotate-90" aria-hidden="true"></i> Pull Requests
-        </a>
-      {{/menu.item}}
-
-      {{#menu.item}}
-        <a href="https://github.com/offirgolan/ember-burger-menu/blob/master/README.md" target="_blank">
-          <i class="fa fa-file-text-o" aria-hidden="true"></i> Documentation
-        </a>
-      {{/menu.item}}
-
-      {{#menu.item}}
-        <a href="https://github.com/offirgolan/ember-burger-menu/graphs" target="_blank">
-          <i class="fa fa-bar-chart" aria-hidden="true"></i> Analytics
-        </a>
-      {{/menu.item}}
-    </ul>
+    {{partial 'menu'}}
   {{/burger.menu}}
 
   {{#burger.outlet}}
@@ -48,21 +18,7 @@
       <i class="fa fa-bars" aria-hidden="true"></i>
     </a>
     <main class="text-center">
-      <div class="header">
-          <img src="images/ember-logo.png" ondragstart="return false;"/>
-          <h1>Burger Menu</h1>
-
-          <div class="description">
-            <p>
-              An off-canvas sidebar component with a collection of animations and styles using CSS transitions.
-            </p>
-          </div>
-
-          <div class="nav">
-            {{#link-to 'index'}}Home{{/link-to}}
-            {{#link-to 'multiple'}}Multiple{{/link-to}}
-          </div>
-      </div>
+      {{partial 'header'}}
 
       <div class="content">
         <div class="container">
@@ -136,11 +92,7 @@
         </div>
         </div>
 
-        <div class="inspiration">
-          <p>
-            Inspired by <a href="https://github.com/codrops/SidebarTransitions">Sidebar Transitions</a> by Codrops
-          </p>
-        </div>
+        {{partial 'inspiration'}}
       </div>
     </main>
   {{/burger.outlet}}
